Add character limit setting
This new option allows you to truncate the file names
to a fixed number of characters.

The expected use case is if you often use this menu to open a directory
with long file names (such as my "papers" folder...),
causing the menu to become exceedingly large:
by setting a character limit, the file name is shortened
and the menu width is also returned to more acceptable margins.