Add DEV suffix and script to remove it
Actually done before the previous commit, but I got an idea to handle it
only in the middle of the work.

The purpose is to have both the cinnamon-spices-applets stable version
and this development version installed at the same time.

Run remove-DEV.py when ready for deployment on the official repo.

Also, either this commit or the previous one updates popup_menu.py
permissions.